UIF-397 Move common parts to common place
