+++ src/common/utils/index.js
@@ -1,3 +1,5 @@
+export * from './exportRolloverErrors';
+export * from './exportRolloverResult';
 export * from './getFiscalYearsForSelect';
 export * from './fundValidation';
 export * from './getExpenseClassesForSelect';
